// Bottleneko.Client/src/lib/api/dtos.gen.ts
export interface DiscordProtocolConfiguration {
    $type: 'Discord';
    token: string;
    receiveEvents: boolean;
    isPresenceIntentEnabled: boolean;
    isServerMembersIntentEnabled: boolean;
    isMessageContentIntentEnabled: boolean;
    proxyId: string | null;
}

export type ProtocolConfiguration
    = DiscordProtocolConfiguration
        | TelegramProtocolConfiguration
        | TwitchProtocolConfiguration;

export interface TelegramProtocolConfiguration {
    $type: 'Telegram';
    token: string;
    receiveEvents: boolean;
    proxyId: string | null;
}
// ...
export interface TwitchAuth {
    clientId: string;
    me: string;
    accessToken: string;
    refreshToken: string;
    scopes: TwitchScope[];
}

export interface TwitchProtocolChannel {
    name: string;
    eventSubscriptions: TwitchSubscriptionTopic[];
}

export interface TwitchProtocolConfiguration {
    $type: 'Twitch';
    auth: TwitchAuth;
    receiveEvents: boolean;
    channels: TwitchProtocolChannel[];
    proxyId: string | null;
}
// ...
export type Protocol
    = 'Discord'
        | 'Telegram'
        | 'Twitch';